5a. Guess the random number
A program that generates a random whole number between 1 and 10 and then asks the user to guess the number. When the guess is not correct, the program tells the user that the guess is too high or too low and then asks the user to try again . The program completes when the user correctly guesses the number.